copilot-todo-agent/requirements.txt
agent-framework>=1.0.0,<2
agent-framework-copilotstudio>=1.0.0b260402
msgraph-sdk>=1.0.0
azure-identity>=1.15.0
python-dotenv>=1.0.0
